requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
click==8.3.1
colorama==0.4.6
fastapi==0.135.1
h11==0.16.0
idna==3.11

kafka-python==2.3.0
confluent-kafka==2.3.0

pydantic==2.12.5
pydantic_core==2.41.5

python-dotenv==1.2.2

starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0

uvicorn==0.42.0

psutil==5.9.8

beautifulsoup4==4.12.3
lxml==5.2.2
requests==2.32.3

# Testing - Phase 3
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-asyncio==0.21.1

httpx==0.28.1